Use 'type=virtual-kubelet' label for node affinity
When setting the node affinity in the base VK deployment manifest, use
'type' instead of 'kubernetes.io/hostname' as the label key. The former
is the a general label for all VK deployments, while hostname is
different for each VK pod.

This is useful for better supporting setups with multiple VK instances
running in the same cluster.