Fix memory calculation for custom instance types
When creating an instance on GCP with `api.ResourceSpec{Memory: "0.5Gi", CPU: "1.0"}`, we get a custom type with 0.9GB memory: `instanceType: "n1-custom-1-921"`.

That comes from the script generating instance data: https://github.com/elotl/kip/blob/master/scripts/create_instance_data/create_gce_instance_data.py#L33

https://cloud.google.com/compute/docs/instances/creating-instance-with-custom-machine-type#n1_custom_machine_types says:

> For N1 machine types, select between 0.9 GB and 6.5 GB per vCPU, inclusive.

but apparently an additional requirement is that it _still_ needs to be a multiple of 0.25GB.

I added another check for the amount of memory when generating instance sizes, ensuring that it is a multiple of the base memory unit (0.25GB).

This fixes #171