Use SSM Parameter Store on AWS to distribute instance cert/key
This is behind a config flag, so we can have users opt in; the default
is still the old method of distributing the certificate and key to
instances via instance user-data.

If enabled, the node controller will create an encrypted parameter with
instance-specific configuration data such as the certificate, key,
cell-config, etc. Itzo-launcher will retrieve it, deserialize the data
and save the files, using the same paths as before.

I also refactored the cloudinitfile package to pull out the logic on
specific files, so it's the same for cloud-init and SSM.

SSM access requires additional IAM permissions. I added them to the TF
example, and also updated docs.