Fix race conditions between instances and security groups deletion in cleanup-vpc.sh
I experienced issues while running this script with Terraform 0.14. Sometime
the VPC wouldn’t get deleted because of a dependant security group. The
theory is that it was a left-over instances that wasn’t completetely
destroyed. Therefor its ENI would prevent the security group from being
deleted, which in turn caused the VPC to get stuck.

This commit does two things:

1. Improve the instances’ termination loop by waiting for the instances
   to be terminated with `aws ec2 wait instance-terminated`.
2. Wrap the security group deletion logic in a retry loop. If at the end of
   the loop there’s any security group left-over, the script will attempt to
   delete it again.